Use int instead of long for docId and long instead of int for n-gram hash. fix occurrences typo and main class in build.sbt
The current maximum docId is 10.103.731 The max value of a signed int is 2.147.483.649 which is more than enough.

There are 2.130.755  articles in the german wikipedia with an average of 1167 words (http://de.wikicount.net/) = ~ 2.486.591.085 words
Let's say 4-grams are to 95% unique, this would mean 2.362.261.530 unique keys will be required. Over/pretty close to the cap, better use a long